Refactor runner helpers into core modules
