+++ flujo/application/runner.py
@@ -4,9 +4,7 @@ import asyncio
 import inspect
 import time
 import weakref
-import contextvars
 import copy
-from unittest.mock import Mock
 from typing import (
     Any,
     Callable,
@@ -16,13 +14,10 @@ from typing import (
     Type,
     TypeVar,
     AsyncIterator,
-    Awaitable,
     Union,
     cast,
-    TypeAlias,
     get_type_hints,
-    get_origin,
-    get_args,
+    Literal,
 )
 
 from pydantic import ValidationError
@@ -34,23 +29,14 @@ from ..exceptions import (
     UsageLimitExceededError,
     PipelineAbortSignal,
     PausedException,
-    MissingAgentError,
     TypeMismatchError,
     ContextInheritanceError,
     InfiniteFallbackError,
+    InfiniteRedirectError as _InfiniteRedirectError,
 )
 from ..domain.dsl.step import Step
 from ..domain.dsl.pipeline import Pipeline
-from ..domain.dsl.loop import LoopStep
-from ..domain.dsl.conditional import ConditionalStep
-from ..domain.dsl.parallel import ParallelStep
-from ..domain.dsl.step import (
-    HumanInTheLoopStep,
-    BranchKey,
-)
-from flujo.steps.cache_step import CacheStep, _generate_cache_key
-from ..domain.plugins import PluginOutcome
-from ..domain.validation import ValidationResult
+from ..domain.dsl.step import HumanInTheLoopStep
 from ..domain.models import (
     BaseModel,
     PipelineResult,
@@ -59,30 +45,21 @@ from ..domain.models import (
     PipelineContext,
     HumanInteraction,
 )
-from pydantic import BaseModel as PydanticBaseModel
 from ..domain.commands import AgentCommand, ExecutedCommandLog
 from pydantic import TypeAdapter
 from ..domain.resources import AppResources
 from ..domain.types import HookCallable
-from ..domain.events import (
-    HookPayload,
-    PreRunPayload,
-    PostRunPayload,
-    PreStepPayload,
-    PostStepPayload,
-    OnStepFailurePayload,
-)
 from ..domain.backends import ExecutionBackend, StepExecutionRequest
 from ..tracing import ConsoleTracer
 
 from .context_manager import (
     _accepts_param,
     _extract_missing_fields,
-    _get_validation_flags,
-    _apply_validation_metadata,
     _types_compatible,
 )
-from .parallel import _execute_parallel_step_logic
+from .core.step_logic import _run_step_logic
+from .core.context_adapter import _build_context_update, _inject_context
+from .core.hook_dispatcher import _dispatch_hook as _dispatch_hook_impl
 
 _signature_cache_weak: weakref.WeakKeyDictionary[Callable[..., Any], inspect.Signature] = (
     weakref.WeakKeyDictionary()
@@ -168,691 +145,16 @@ def _cached_type_hints(func: Callable[..., Any]) -> Dict[str, Any] | None:
     return hints
 
 
-_fallback_chain_var: contextvars.ContextVar[list[Step[Any, Any]]] = contextvars.ContextVar(
-    "_fallback_chain", default=[]
-)
-
 _agent_command_adapter: TypeAdapter[AgentCommand] = TypeAdapter(AgentCommand)
 
 
-class InfiniteRedirectError(OrchestratorError):
-    """Raised when a redirect loop is detected."""
+# Alias exported for backwards compatibility
+InfiniteRedirectError = _InfiniteRedirectError
 
 
 RunnerInT = TypeVar("RunnerInT")
 RunnerOutT = TypeVar("RunnerOutT")
 ContextT = TypeVar("ContextT", bound=BaseModel)
-
-StepExecutor: TypeAlias = Callable[
-    [Step[Any, Any], Any, Optional[ContextT], Optional[AppResources]],
-    Awaitable[StepResult],
-]
-
-
-async def _execute_loop_step_logic(
-    loop_step: LoopStep[ContextT],
-    loop_step_initial_input: Any,
-    context: Optional[ContextT],
-    resources: Optional[AppResources],
-    *,
-    step_executor: StepExecutor[ContextT],
-    context_model_defined: bool,
-    usage_limits: UsageLimits | None = None,
-) -> StepResult:
-    """Logic for executing a LoopStep without engine coupling."""
-    loop_overall_result = StepResult(name=loop_step.name)
-
-    if loop_step.initial_input_to_loop_body_mapper:
-        try:
-            current_body_input = loop_step.initial_input_to_loop_body_mapper(
-                loop_step_initial_input, context
-            )
-        except Exception as e:
-            telemetry.logfire.error(
-                f"Error in initial_input_to_loop_body_mapper for LoopStep '{loop_step.name}': {e}"
-            )
-            loop_overall_result.success = False
-            loop_overall_result.feedback = f"Initial input mapper raised an exception: {e}"
-            return loop_overall_result
-    else:
-        current_body_input = loop_step_initial_input
-
-    last_successful_iteration_body_output: Any = None
-    final_body_output_of_last_iteration: Any = None
-    loop_exited_successfully_by_condition = False
-
-    for i in range(1, loop_step.max_loops + 1):
-        loop_overall_result.attempts = i
-        telemetry.logfire.info(
-            f"LoopStep '{loop_step.name}': Starting Iteration {i}/{loop_step.max_loops}"
-        )
-
-        iteration_succeeded_fully = True
-        current_iteration_data_for_body_step = current_body_input
-
-        with telemetry.logfire.span(f"Loop '{loop_step.name}' - Iteration {i}"):
-            for body_s in loop_step.loop_body_pipeline.steps:
-                body_step_result_obj = await step_executor(
-                    body_s,
-                    current_iteration_data_for_body_step,
-                    context,
-                    resources,
-                )
-
-                loop_overall_result.latency_s += body_step_result_obj.latency_s
-                loop_overall_result.cost_usd += getattr(body_step_result_obj, "cost_usd", 0.0)
-                loop_overall_result.token_counts += getattr(body_step_result_obj, "token_counts", 0)
-
-                if usage_limits is not None:
-                    if (
-                        usage_limits.total_cost_usd_limit is not None
-                        and loop_overall_result.cost_usd > usage_limits.total_cost_usd_limit
-                    ):
-                        telemetry.logfire.warn(
-                            f"Cost limit of ${usage_limits.total_cost_usd_limit} exceeded"
-                        )
-                        loop_overall_result.success = False
-                        loop_overall_result.feedback = (
-                            f"Cost limit of ${usage_limits.total_cost_usd_limit} exceeded"
-                        )
-                        pr: PipelineResult[ContextT] = PipelineResult(
-                            step_history=[loop_overall_result],
-                            total_cost_usd=loop_overall_result.cost_usd,
-                        )
-                        Flujo._set_final_context(pr, context)
-                        raise UsageLimitExceededError(
-                            loop_overall_result.feedback,
-                            pr,
-                        )
-                    if (
-                        usage_limits.total_tokens_limit is not None
-                        and loop_overall_result.token_counts > usage_limits.total_tokens_limit
-                    ):
-                        telemetry.logfire.warn(
-                            f"Token limit of {usage_limits.total_tokens_limit} exceeded"
-                        )
-                        loop_overall_result.success = False
-                        loop_overall_result.feedback = (
-                            f"Token limit of {usage_limits.total_tokens_limit} exceeded"
-                        )
-                        pr_tokens: PipelineResult[ContextT] = PipelineResult(
-                            step_history=[loop_overall_result],
-                            total_cost_usd=loop_overall_result.cost_usd,
-                        )
-                        Flujo._set_final_context(pr_tokens, context)
-                        raise UsageLimitExceededError(
-                            loop_overall_result.feedback,
-                            pr_tokens,
-                        )
-
-                if not body_step_result_obj.success:
-                    telemetry.logfire.warn(
-                        f"Body Step '{body_s.name}' in LoopStep '{loop_step.name}' (Iteration {i}) failed."
-                    )
-                    iteration_succeeded_fully = False
-                    final_body_output_of_last_iteration = body_step_result_obj.output
-                    break
-
-                current_iteration_data_for_body_step = body_step_result_obj.output
-
-        if iteration_succeeded_fully:
-            last_successful_iteration_body_output = current_iteration_data_for_body_step
-        final_body_output_of_last_iteration = current_iteration_data_for_body_step
-
-        try:
-            should_exit = loop_step.exit_condition_callable(
-                final_body_output_of_last_iteration, context
-            )
-        except Exception as e:
-            telemetry.logfire.error(
-                f"Error in exit_condition_callable for LoopStep '{loop_step.name}': {e}"
-            )
-            loop_overall_result.success = False
-            loop_overall_result.feedback = f"Exit condition callable raised an exception: {e}"
-            break
-
-        if should_exit:
-            telemetry.logfire.info(
-                f"LoopStep '{loop_step.name}' exit condition met at iteration {i}."
-            )
-            loop_overall_result.success = iteration_succeeded_fully
-            if not iteration_succeeded_fully:
-                loop_overall_result.feedback = (
-                    "Loop exited by condition, but last iteration body failed."
-                )
-            loop_exited_successfully_by_condition = True
-            break
-
-        if i < loop_step.max_loops:
-            if loop_step.iteration_input_mapper:
-                try:
-                    current_body_input = loop_step.iteration_input_mapper(
-                        final_body_output_of_last_iteration, context, i
-                    )
-                except Exception as e:
-                    telemetry.logfire.error(
-                        f"Error in iteration_input_mapper for LoopStep '{loop_step.name}': {e}"
-                    )
-                    loop_overall_result.success = False
-                    loop_overall_result.feedback = (
-                        f"Iteration input mapper raised an exception: {e}"
-                    )
-                    break
-            else:
-                current_body_input = final_body_output_of_last_iteration
-    else:
-        telemetry.logfire.warn(
-            f"LoopStep '{loop_step.name}' reached max_loops ({loop_step.max_loops}) without exit condition being met."
-        )
-        loop_overall_result.success = False
-        loop_overall_result.feedback = (
-            f"Reached max_loops ({loop_step.max_loops}) without meeting exit condition."
-        )
-
-    if loop_overall_result.success and loop_exited_successfully_by_condition:
-        if loop_step.loop_output_mapper:
-            try:
-                loop_overall_result.output = loop_step.loop_output_mapper(
-                    last_successful_iteration_body_output, context
-                )
-            except Exception as e:
-                telemetry.logfire.error(
-                    f"Error in loop_output_mapper for LoopStep '{loop_step.name}': {e}"
-                )
-                loop_overall_result.success = False
-                loop_overall_result.feedback = f"Loop output mapper raised an exception: {e}"
-                loop_overall_result.output = None
-        else:
-            loop_overall_result.output = last_successful_iteration_body_output
-    else:
-        loop_overall_result.output = final_body_output_of_last_iteration
-        if not loop_overall_result.feedback:
-            loop_overall_result.feedback = (
-                "Loop did not complete successfully or exit condition not met positively."
-            )
-
-    return loop_overall_result
-
-
-async def _execute_conditional_step_logic(
-    conditional_step: ConditionalStep[ContextT],
-    conditional_step_input: Any,
-    context: Optional[ContextT],
-    resources: Optional[AppResources],
-    *,
-    step_executor: StepExecutor[ContextT],
-    context_model_defined: bool,
-    usage_limits: UsageLimits | None = None,
-) -> StepResult:
-    """Logic for executing a ConditionalStep without engine coupling."""
-    conditional_overall_result = StepResult(name=conditional_step.name)
-    executed_branch_key: BranchKey | None = None
-    branch_output: Any = None
-    branch_succeeded = False
-
-    try:
-        branch_key_to_execute = conditional_step.condition_callable(conditional_step_input, context)
-        telemetry.logfire.info(
-            f"ConditionalStep '{conditional_step.name}': Condition evaluated to branch key '{branch_key_to_execute}'."
-        )
-        executed_branch_key = branch_key_to_execute
-
-        selected_branch_pipeline = conditional_step.branches.get(branch_key_to_execute)
-        if selected_branch_pipeline is None:
-            selected_branch_pipeline = conditional_step.default_branch_pipeline
-            if selected_branch_pipeline is None:
-                err_msg = f"ConditionalStep '{conditional_step.name}': No branch found for key '{branch_key_to_execute}' and no default branch defined."
-                telemetry.logfire.warn(err_msg)
-                conditional_overall_result.success = False
-                conditional_overall_result.feedback = err_msg
-                return conditional_overall_result
-            telemetry.logfire.info(
-                f"ConditionalStep '{conditional_step.name}': Executing default branch."
-            )
-        else:
-            telemetry.logfire.info(
-                f"ConditionalStep '{conditional_step.name}': Executing branch for key '{branch_key_to_execute}'."
-            )
-
-        if conditional_step.branch_input_mapper:
-            input_for_branch = conditional_step.branch_input_mapper(conditional_step_input, context)
-        else:
-            input_for_branch = conditional_step_input
-
-        current_branch_data = input_for_branch
-        branch_pipeline_failed_internally = False
-
-        for branch_s in selected_branch_pipeline.steps:
-            with telemetry.logfire.span(
-                f"ConditionalStep '{conditional_step.name}' Branch '{branch_key_to_execute}' - Step '{branch_s.name}'"
-            ) as span:
-                if executed_branch_key is not None:
-                    try:
-                        span.set_attribute("executed_branch_key", str(executed_branch_key))
-                    except Exception as e:  # pragma: no cover - defensive
-                        telemetry.logfire.error(f"Error setting span attribute: {e}")
-                branch_step_result_obj = await step_executor(
-                    branch_s,
-                    current_branch_data,
-                    context,
-                    resources,
-                )
-
-            conditional_overall_result.latency_s += branch_step_result_obj.latency_s
-            conditional_overall_result.cost_usd += getattr(branch_step_result_obj, "cost_usd", 0.0)
-            conditional_overall_result.token_counts += getattr(
-                branch_step_result_obj, "token_counts", 0
-            )
-
-            if not branch_step_result_obj.success:
-                telemetry.logfire.warn(
-                    f"Step '{branch_s.name}' in branch '{branch_key_to_execute}' of ConditionalStep '{conditional_step.name}' failed."
-                )
-                branch_pipeline_failed_internally = True
-                branch_output = branch_step_result_obj.output
-                conditional_overall_result.feedback = f"Failure in branch '{branch_key_to_execute}', step '{branch_s.name}': {branch_step_result_obj.feedback}"
-                break
-
-            current_branch_data = branch_step_result_obj.output
-
-        if not branch_pipeline_failed_internally:
-            branch_output = current_branch_data
-            branch_succeeded = True
-
-    except Exception as e:
-        telemetry.logfire.error(
-            f"Error during ConditionalStep '{conditional_step.name}' execution: {e}",
-            exc_info=True,
-        )
-        conditional_overall_result.success = False
-        conditional_overall_result.feedback = f"Error executing conditional logic or branch: {e}"
-        return conditional_overall_result
-
-    conditional_overall_result.success = branch_succeeded
-    if branch_succeeded:
-        if conditional_step.branch_output_mapper:
-            try:
-                conditional_overall_result.output = conditional_step.branch_output_mapper(
-                    branch_output, executed_branch_key, context
-                )
-            except Exception as e:
-                telemetry.logfire.error(
-                    f"Error in branch_output_mapper for ConditionalStep '{conditional_step.name}': {e}"
-                )
-                conditional_overall_result.success = False
-                conditional_overall_result.feedback = (
-                    f"Branch output mapper raised an exception: {e}"
-                )
-                conditional_overall_result.output = None
-        else:
-            conditional_overall_result.output = branch_output
-    else:
-        conditional_overall_result.output = branch_output
-
-    conditional_overall_result.attempts = 1
-    if executed_branch_key is not None:
-        conditional_overall_result.metadata_ = conditional_overall_result.metadata_ or {}
-        conditional_overall_result.metadata_["executed_branch_key"] = str(executed_branch_key)
-
-    return conditional_overall_result
-
-
-async def _run_step_logic(
-    step: Step[Any, Any],
-    data: Any,
-    context: Optional[ContextT],
-    resources: Optional[AppResources],
-    *,
-    step_executor: StepExecutor[ContextT],
-    context_model_defined: bool,
-    usage_limits: UsageLimits | None = None,
-) -> StepResult:
-    """Core logic for executing a single step without engine coupling."""
-    visited_ids: set[int] = set()
-    if step.agent is not None:
-        visited_ids.add(id(step.agent))
-    if isinstance(step, CacheStep):
-        key = _generate_cache_key(step.wrapped_step, data, context=context, resources=resources)
-        cached: StepResult | None = None
-        if key:
-            try:
-                cached = await step.cache_backend.get(key)
-            except Exception as e:  # pragma: no cover - defensive
-                telemetry.logfire.warn(f"Cache get failed for key {key}: {e}")
-        if isinstance(cached, StepResult):
-            result = cached.model_copy(deep=True)
-            result.metadata_ = result.metadata_ or {}
-            result.metadata_["cache_hit"] = True
-            return result
-
-        result = await step_executor(step.wrapped_step, data, context, resources)
-        if result.success and key:
-            try:
-                await step.cache_backend.set(key, result)
-            except Exception as e:  # pragma: no cover - defensive
-                telemetry.logfire.warn(f"Cache set failed for key {key}: {e}")
-        return result
-    if isinstance(step, LoopStep):
-        return await _execute_loop_step_logic(
-            step,
-            data,
-            context,
-            resources,
-            step_executor=step_executor,
-            context_model_defined=context_model_defined,
-            usage_limits=usage_limits,
-        )
-    if isinstance(step, ConditionalStep):
-        return await _execute_conditional_step_logic(
-            step,
-            data,
-            context,
-            resources,
-            step_executor=step_executor,
-            context_model_defined=context_model_defined,
-            usage_limits=usage_limits,
-        )
-    if isinstance(step, ParallelStep):
-        return await _execute_parallel_step_logic(
-            step,
-            data,
-            context,
-            resources,
-            step_executor=step_executor,
-            context_model_defined=context_model_defined,
-            usage_limits=usage_limits,
-            context_setter=Flujo._set_final_context,
-        )
-    if isinstance(step, HumanInTheLoopStep):
-        message = step.message_for_user if step.message_for_user is not None else str(data)
-        if isinstance(context, PipelineContext):
-            context.scratchpad["status"] = "paused"
-        raise PausedException(message)
-
-    result = StepResult(name=step.name)
-    original_agent = step.agent
-    current_agent = original_agent
-    last_feedback = None
-    last_raw_output = None
-    last_unpacked_output = None
-    validation_failed = False
-    last_attempt_feedbacks: list[str] = []
-    last_attempt_output = None
-    for attempt in range(1, step.config.max_retries + 1):
-        validation_failed = False
-        result.attempts = attempt
-        feedbacks: list[str] = []  # feedbacks for this attempt only
-        plugin_failed_this_attempt = False  # Always initialize at start of attempt
-        if current_agent is None:
-            raise MissingAgentError(
-                f"Step '{step.name}' is missing an agent. Assign one via `Step('name', agent=...)` "
-                "or by using a step factory like `@step` or `Step.from_callable()`."
-            )
-
-        start = time.monotonic()
-        agent_kwargs: Dict[str, Any] = {}
-        # Apply prompt processors
-        if step.processors.prompt_processors:
-            telemetry.logfire.info(
-                f"Running {len(step.processors.prompt_processors)} prompt processors for step '{step.name}'..."
-            )
-            processed = data
-            for proc in step.processors.prompt_processors:
-                try:
-                    processed = await proc.process(processed, context)
-                except Exception as e:  # pragma: no cover - defensive
-                    telemetry.logfire.error(f"Processor {proc.name} failed: {e}")
-                data = processed
-        from ..signature_tools import analyze_signature
-
-        target = getattr(current_agent, "_agent", current_agent)
-        func = getattr(target, "_step_callable", target.run)
-        spec = analyze_signature(func)
-
-        if spec.needs_context:
-            if context is None:
-                raise TypeError(
-                    f"Component in step '{step.name}' requires a context, but no context model was provided to the Flujo runner."
-                )
-            agent_kwargs["context"] = context
-
-        if resources is not None:
-            if spec.needs_resources:
-                agent_kwargs["resources"] = resources
-            elif _accepts_param(func, "resources"):
-                agent_kwargs["resources"] = resources
-        if step.config.temperature is not None and _accepts_param(func, "temperature"):
-            agent_kwargs["temperature"] = step.config.temperature
-        raw_output = await current_agent.run(data, **agent_kwargs)
-        result.latency_s += time.monotonic() - start
-        last_raw_output = raw_output
-
-        if isinstance(raw_output, Mock):
-            raise TypeError(
-                f"Step '{step.name}' returned a Mock object. This is usually due to "
-                "an unconfigured mock in a test. Please configure your mock agent "
-                "to return a concrete value."
-            )
-
-        unpacked_output = getattr(raw_output, "output", raw_output)
-        # Apply output processors
-        if step.processors.output_processors:
-            telemetry.logfire.info(
-                f"Running {len(step.processors.output_processors)} output processors for step '{step.name}'..."
-            )
-            processed = unpacked_output
-            for proc in step.processors.output_processors:
-                try:
-                    processed = await proc.process(processed, context)
-                except Exception as e:  # pragma: no cover - defensive
-                    telemetry.logfire.error(f"Processor {proc.name} failed: {e}")
-                unpacked_output = processed
-        last_unpacked_output = unpacked_output
-
-        success = True
-        redirect_to = None
-        final_plugin_outcome: PluginOutcome | None = None
-        is_validation_step, is_strict = _get_validation_flags(step)
-
-        sorted_plugins = sorted(step.plugins, key=lambda p: p[1], reverse=True)
-        for plugin, _ in sorted_plugins:
-            try:
-                from ..signature_tools import analyze_signature
-
-                plugin_kwargs: Dict[str, Any] = {}
-                func = getattr(plugin, "_plugin_callable", plugin.validate)
-                spec = analyze_signature(func)
-
-                if spec.needs_context:
-                    if context is None:
-                        raise TypeError(
-                            f"Plugin in step '{step.name}' requires a context, but no context model was provided to the Flujo runner."
-                        )
-                    plugin_kwargs["context"] = context
-
-                if resources is not None:
-                    if spec.needs_resources:
-                        plugin_kwargs["resources"] = resources
-                    elif _accepts_param(func, "resources"):
-                        plugin_kwargs["resources"] = resources
-                validated = await asyncio.wait_for(
-                    plugin.validate(
-                        {"output": last_unpacked_output, "feedback": last_feedback},
-                        **plugin_kwargs,
-                    ),
-                    timeout=step.config.timeout_s,
-                )
-            except asyncio.TimeoutError as e:
-                raise TimeoutError(f"Plugin timeout in step {step.name}") from e
-
-            if not validated.success:
-                validation_failed = True
-                plugin_failed_this_attempt = True
-                if validated.feedback:
-                    feedbacks.append(validated.feedback)
-                redirect_to = validated.redirect_to
-                final_plugin_outcome = validated
-            if validated.new_solution is not None:
-                final_plugin_outcome = validated
-
-        if final_plugin_outcome and final_plugin_outcome.new_solution is not None:
-            unpacked_output = final_plugin_outcome.new_solution
-            last_unpacked_output = unpacked_output
-
-        # Run programmatic validators regardless of plugin outcome
-        if step.validators:
-            telemetry.logfire.info(
-                f"Running {len(step.validators)} programmatic validators for step '{step.name}'..."
-            )
-            validation_tasks = [
-                validator.validate(unpacked_output, context=context)
-                for validator in step.validators
-            ]
-            try:
-                validation_results = await asyncio.gather(*validation_tasks, return_exceptions=True)
-            except Exception as e:  # pragma: no cover - defensive
-                validation_results = [e]
-
-            failed_checks_feedback: list[str] = []
-            collected_results: list[ValidationResult] = []
-            for validator, res in zip(step.validators, validation_results):
-                if isinstance(res, Exception):
-                    vname = getattr(
-                        validator,
-                        "name",
-                        getattr(validator, "__class__", type(validator)).__name__,
-                    )
-                    failed_checks_feedback.append(f"Validator '{vname}' crashed: {res}")
-                    continue
-                vres = cast(ValidationResult, res)
-                collected_results.append(vres)
-                if not vres.is_valid:
-                    fb = vres.feedback or "No details provided."
-                    failed_checks_feedback.append(f"Check '{vres.validator_name}' failed: {fb}")
-
-            if step.persist_validation_results_to and context is not None:
-                if hasattr(context, step.persist_validation_results_to):
-                    history_list = getattr(context, step.persist_validation_results_to)
-                    if isinstance(history_list, list):
-                        history_list.extend(collected_results)
-
-            if failed_checks_feedback:
-                validation_failed = True
-                feedbacks.extend(failed_checks_feedback)
-                # For non-strict validation steps, don't fail the step when validation fails
-                if is_strict or not is_validation_step:
-                    success = False
-
-        # --- RETRY LOGIC FIX ---
-        if plugin_failed_this_attempt:
-            success = False
-        # --- END FIX ---
-        # --- JOIN ALL FEEDBACKS ---
-        feedback = "\n".join(feedbacks).strip() if feedbacks else None
-        # --- END JOIN ---
-        if not success and attempt == step.config.max_retries:
-            last_attempt_feedbacks = feedbacks.copy()
-            last_attempt_output = last_unpacked_output
-        if success:
-            result.output = unpacked_output
-            result.success = True
-            result.feedback = feedback
-            result.token_counts += getattr(raw_output, "token_counts", 0)
-            result.cost_usd += getattr(raw_output, "cost_usd", 0.0)
-            _apply_validation_metadata(
-                result,
-                validation_failed=validation_failed,
-                is_validation_step=is_validation_step,
-                is_strict=is_strict,
-            )
-            return result
-
-        for handler in step.failure_handlers:
-            handler()
-
-        if redirect_to:
-            redirect_id = id(redirect_to)
-            if redirect_id in visited_ids:
-                raise InfiniteRedirectError(f"Redirect loop detected in step {step.name}")
-            visited_ids.add(redirect_id)
-            current_agent = redirect_to
-        else:
-            current_agent = original_agent
-
-        if feedback:
-            if isinstance(data, dict):
-                data["feedback"] = data.get("feedback", "") + "\n" + feedback
-            else:
-                data = f"{str(data)}\n{feedback}"
-        last_feedback = feedback
-
-    # After all retries, set feedback to last attempt's feedbacks
-    result.success = False
-    result.feedback = (
-        "\n".join(last_attempt_feedbacks).strip() if last_attempt_feedbacks else last_feedback
-    )
-    is_validation_step, is_strict = _get_validation_flags(step)
-    if validation_failed and is_strict:
-        result.output = None
-    else:
-        result.output = last_attempt_output
-    result.token_counts += (
-        getattr(last_raw_output, "token_counts", 1) if last_raw_output is not None else 0
-    )
-    result.cost_usd += (
-        getattr(last_raw_output, "cost_usd", 0.0) if last_raw_output is not None else 0.0
-    )
-    _apply_validation_metadata(
-        result,
-        validation_failed=validation_failed,
-        is_validation_step=is_validation_step,
-        is_strict=is_strict,
-    )
-    # If the step failed and a fallback is defined, execute it.
-    if not result.success and step.fallback_step:
-        telemetry.logfire.info(
-            f"Step '{step.name}' failed. Attempting fallback step '{step.fallback_step.name}'."
-        )
-        original_failure_feedback = result.feedback
-
-        chain = _fallback_chain_var.get()
-        if step in chain:
-            raise InfiniteFallbackError(f"Fallback loop detected in step '{step.name}'")
-        token = _fallback_chain_var.set(chain + [step])
-        try:
-            fallback_result = await step_executor(
-                step.fallback_step,
-                data,
-                context,
-                resources,
-            )
-        finally:
-            _fallback_chain_var.reset(token)
-
-        result.latency_s += fallback_result.latency_s
-        result.cost_usd += fallback_result.cost_usd
-        result.token_counts += fallback_result.token_counts
-
-        if fallback_result.success:
-            result.success = True
-            result.output = fallback_result.output
-            result.feedback = None
-            result.metadata_ = {
-                **(result.metadata_ or {}),
-                "fallback_triggered": True,
-                "original_error": original_failure_feedback,
-            }
-        else:
-            result.feedback = (
-                f"Original error: {original_failure_feedback}\n"
-                f"Fallback error: {fallback_result.feedback}"
-            )
-    if not result.success and step.persist_feedback_to_context:
-        if context is not None and hasattr(context, step.persist_feedback_to_context):
-            history_list = getattr(context, step.persist_feedback_to_context)
-            if isinstance(history_list, list) and result.feedback:
-                history_list.append(result.feedback)
-    return result
 
 
 class Flujo(Generic[RunnerInT, RunnerOutT, ContextT]):
@@ -890,57 +192,20 @@ class Flujo(Generic[RunnerInT, RunnerOutT, ContextT]):
             backend = LocalBackend()
         self.backend = backend
 
-    async def _dispatch_hook(self, event_name: str, **kwargs: Any) -> None:
-        """Invoke registered hooks with type-based filtering.
+    async def _dispatch_hook(
+        self,
+        event_name: Literal[
+            "pre_run",
+            "post_run",
+            "pre_step",
+            "post_step",
+            "on_step_failure",
+        ],
+        **kwargs: Any,
+    ) -> None:
+        """Invoke registered hooks for ``event_name``."""
 
-        The runner supports a simple event system that allows external code to
-        observe pipeline execution. Hooks are plain callables that accept a
-        subclass of :class:`HookPayload`. To avoid unnecessary errors we inspect
-        the hook's first parameter annotation and only call it if it is
-        compatible with the payload type for the current ``event_name``.
-        """
-
-        payload_map: dict[str, type[HookPayload]] = {
-            "pre_run": PreRunPayload,
-            "post_run": PostRunPayload,
-            "pre_step": PreStepPayload,
-            "post_step": PostStepPayload,
-            "on_step_failure": OnStepFailurePayload,
-        }
-        PayloadCls = payload_map.get(event_name)
-        if PayloadCls is None:
-            return
-
-        payload = PayloadCls(event_name=cast(Any, event_name), **kwargs)
-
-        for hook in self.hooks:
-            try:
-                should_call = True
-                try:
-                    sig = _cached_signature(hook)
-                    params = list(sig.parameters.values()) if sig else []
-                    if params:
-                        hints = _cached_type_hints(hook) or {}
-                        ann = hints.get(params[0].name, params[0].annotation)
-                        if ann is not inspect.Signature.empty:
-                            origin = get_origin(ann)
-                            if origin is Union:
-                                if not any(isinstance(payload, t) for t in get_args(ann)):
-                                    should_call = False
-                            elif isinstance(ann, type):
-                                if not isinstance(payload, ann):
-                                    should_call = False
-                except Exception as e:
-                    name = getattr(hook, "__name__", str(hook))
-                    telemetry.logfire.error(f"Error in hook '{name}': {e}")
-
-                if should_call:
-                    await hook(payload)
-            except PipelineAbortSignal:
-                raise
-            except Exception as e:
-                name = getattr(hook, "__name__", str(hook))
-                telemetry.logfire.error(f"Error in hook '{name}': {e}")
+        await _dispatch_hook_impl(self.hooks, event_name, **kwargs)
 
     async def _run_step(
         self,
@@ -985,31 +250,19 @@ class Flujo(Generic[RunnerInT, RunnerOutT, ContextT]):
         result = await self.backend.execute_step(request)
         if getattr(step, "updates_context", False):
             if self.context_model is not None and context is not None:
-                if isinstance(result.output, (BaseModel, PydanticBaseModel)):
-                    update_data = result.output.model_dump(exclude_unset=True)
-                elif isinstance(result.output, dict):
-                    update_data = result.output
-                else:
+                update_data = _build_context_update(result.output)
+                if update_data is None:
                     telemetry.logfire.warn(
                         f"Step '{step.name}' has updates_context=True but did not return a dict or Pydantic model. "
                         "Skipping context update."
                     )
                     return result
 
-                try:
-                    original_data = context.model_dump()
-                    for key, value in update_data.items():
-                        setattr(context, key, value)
-
-                    validated = self.context_model.model_validate(context.model_dump())
-                    context.__dict__.update(validated.__dict__)
-                except ValidationError as e:
-                    for key, value in original_data.items():
-                        setattr(context, key, value)
+                err = _inject_context(context, update_data, self.context_model)
+                if err is not None:
                     error_msg = (
-                        f"Context update by step '{step.name}' failed Pydantic validation: {e}"
+                        f"Context update by step '{step.name}' failed Pydantic validation: {err}"
                     )
-                    telemetry.logfire.error(error_msg)
                     result.success = False
                     result.feedback = error_msg
                     return result
